Consolidate Row & Inline
