+++ src/design-system/components/Inline/Inline.examples.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable sort-keys-fix/sort-keys-fix */
 import React from 'react';
+import { View } from 'react-native';
 import { Example } from '../../docs/types';
 import source from '../../docs/utils/source.macro';
 import { Placeholder } from '../../playground/Placeholder';
@@ -10,6 +11,25 @@ export const basicUsage: Example = {
   Example: () =>
     source(
       <Inline space="12px">
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+        <Placeholder height={40} width={40} />
+      </Inline>
+    ),
+};
+
+export const noWrap: Example = {
+  name: 'No wrap',
+  Example: () =>
+    source(
+      <Inline space="12px" wrap={false}>
         <Placeholder height={40} width={40} />
         <Placeholder height={40} width={40} />
         <Placeholder height={40} width={40} />
@@ -115,6 +135,18 @@ export const rightAlignedHorizontally: Example = {
     ),
 };
 
+export const justifiedHorizontally: Example = {
+  name: 'Justified horizontally',
+  Example: () =>
+    source(
+      <Inline alignHorizontal="justify" space="19px">
+        <Placeholder width={40} />
+        <Placeholder width={40} />
+        <Placeholder width={40} />
+      </Inline>
+    ),
+};
+
 export const centerAlignedVertically: Example = {
   name: 'Center-aligned vertically',
   Example: () =>
@@ -165,6 +197,99 @@ export const centerAlignedHorizontallyVertically: Example = {
         <Placeholder height={70} width={40} />
         <Placeholder height={10} width={40} />
         <Placeholder height={50} width={40} />
+      </Inline>
+    ),
+};
+
+export const fixedHeightSeparators: Example = {
+  name: 'With fixed-height separators',
+  Example: () =>
+    source(
+      <Inline
+        separator={
+          <View style={{ backgroundColor: '#999', height: 20, width: 1 }} />
+        }
+        space="19px"
+        wrap={false}
+      >
+        <Placeholder width={40} />
+        <Placeholder width={40} />
+        <Placeholder width={40} />
+      </Inline>
+    ),
+};
+
+export const fixedHeightSeparatorsVerticalAlignment: Example = {
+  name: 'With fixed-height separators and vertical alignment',
+  Example: () =>
+    source(
+      <Inline
+        alignVertical="center"
+        separator={
+          <View style={{ backgroundColor: '#999', height: 20, width: 1 }} />
+        }
+        space="19px"
+        wrap={false}
+      >
+        <Placeholder height={40} width={40} />
+        <Placeholder height={60} width={40} />
+        <Placeholder height={20} width={40} />
+      </Inline>
+    ),
+};
+
+export const fixedHeightSeparatorsHorizontalAlignment: Example = {
+  name: 'With fixed-height separators and horizontal alignment',
+  Example: () =>
+    source(
+      <Inline
+        alignHorizontal="center"
+        separator={
+          <View style={{ backgroundColor: '#999', height: 20, width: 1 }} />
+        }
+        space="19px"
+        wrap={false}
+      >
+        <Placeholder width={40} />
+        <Placeholder width={40} />
+        <Placeholder width={40} />
+      </Inline>
+    ),
+};
+
+export const fullHeightSeparators: Example = {
+  name: 'With full-height separators',
+  Example: () =>
+    source(
+      <Inline
+        alignVertical="center"
+        separator={
+          <View style={{ backgroundColor: '#999', flexGrow: 1, width: 1 }} />
+        }
+        space="19px"
+        wrap={false}
+      >
+        <Placeholder height={40} width={40} />
+        <Placeholder height={60} width={40} />
+        <Placeholder height={20} width={40} />
+      </Inline>
+    ),
+};
+
+export const noSpaceAndSeparators: Example = {
+  name: 'With no space and separators',
+  Example: () =>
+    source(
+      <Inline
+        alignVertical="center"
+        separator={
+          <View style={{ backgroundColor: '#999', height: 20, width: 1 }} />
+        }
+        wrap={false}
+      >
+        <Placeholder width={40} />
+        <Placeholder width={40} />
+        <Placeholder width={40} />
       </Inline>
     ),
 };
