+++ src/design-system/components/Inline/Inline.playground.tsx
@@ -7,14 +7,21 @@ const playground: Playground = {
   meta,
   examples: [
     examples.basicUsage,
+    examples.noWrap,
     examples.customSpace,
     examples.differentSpaceOnAxis,
     examples.customSpaceOnAxis,
     examples.centerAlignedHorizontally,
     examples.rightAlignedHorizontally,
+    examples.justifiedHorizontally,
     examples.centerAlignedVertically,
     examples.bottomAlignedVertically,
     examples.centerAlignedHorizontallyVertically,
+    examples.fixedHeightSeparators,
+    examples.fixedHeightSeparatorsHorizontalAlignment,
+    examples.fixedHeightSeparatorsVerticalAlignment,
+    examples.fullHeightSeparators,
+    examples.noSpaceAndSeparators,
   ],
 };
