--- src/design-system/components/Row/Row.meta.ts
@@ -1,8 +1,0 @@
-import { Meta } from '../../docs/types';
-
-const meta: Meta = {
-  category: 'Layout',
-  name: 'Row',
-};
-
-export default meta;
