+++ src/design-system/index.ts
@@ -12,7 +12,6 @@ export { Inline } from './components/Inline/Inline';
 export { Inset } from './components/Inset/Inset';
 export { MarkdownText } from './components/MarkdownText/MarkdownText';
 export { renderStringWithEmoji } from './typography/renderStringWithEmoji';
-export { Row } from './components/Row/Row';
 export { Stack } from './components/Stack/Stack';
 export { Text } from './components/Text/Text';
 export { TextLink } from './components/TextLink/TextLink';
@@ -30,7 +29,6 @@ export type { HeadingProps } from './components/Heading/Heading';
 export type { InlineProps } from './components/Inline/Inline';
 export type { InsetProps } from './components/Inset/Inset';
 export type { MarkdownTextProps } from './components/MarkdownText/MarkdownText';
-export type { RowProps } from './components/Row/Row';
 export type { Space } from './layout/space';
 export type { StackProps } from './components/Stack/Stack';
 export type { TextLinkProps } from './components/TextLink/TextLink';
