+++ src/design-system/playground/Playground.tsx
@@ -28,7 +28,6 @@ import headingPlayground from '../components/Heading/Heading.playground';
 import inlinePlayground from '../components/Inline/Inline.playground';
 import insetPlayground from '../components/Inset/Inset.playground';
 import markdownTextPlayground from '../components/MarkdownText/MarkdownText.playground';
-import rowPlayground from '../components/Row/Row.playground';
 import stackPlayground from '../components/Stack/Stack.playground';
 import textPlayground from '../components/Text/Text.playground';
 import textLinkPlayground from '../components/TextLink/TextLink.playground';
@@ -47,7 +46,6 @@ const allDocs = [
   inlinePlayground,
   insetPlayground,
   markdownTextPlayground,
-  rowPlayground,
   stackPlayground,
   textPlayground,
   textLinkPlayground,
